fixed blader level sync and aiger armor model,textures are TO-DO yet

+++ src/main/java/com/grillo78/BeyCraft/capabilities/BladerLevelStorage.java
@@ -16,7 +16,7 @@ public class BladerLevelStorage implements IStorage<IBladerLevel> {
 
 	@Override
 	public void readNBT(Capability<IBladerLevel> capability, IBladerLevel instance, EnumFacing side, NBTBase nbt) {
-		instance.setBladerLevel(((NBTPrimitive) nbt).getFloat());
+		instance.setBladerLevel(((NBTPrimitive) nbt).getInt());
 	}
 
 }
